Compute bearings of the boat with GPS positions.
Previous calculation of bearing based on position was wrong!!! Removed the manual trigonometrical computation of the bearing and computed it with the bearing function from package.

Bearings are calculated per sec over the period with the GPGGA GPS and 15s during transects 3, 4 and 5. To smooth the bearing and remove the noise in GPS position, the bearing are computed between each pair of neighboring positions, then average every 1 min, and finally interpolated on a regular grid of 1s over all transects (with linear interpolation).